Remove unused code from openapi-utilities
